% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.tests.R
\name{generalized.ttest.precalc}
\alias{generalized.ttest.precalc}
\title{Perform a Generalized t-test Using Precalculated Statistics}
\usage{
generalized.ttest.precalc(m, se, df, epsilon_l, epsilon_u)
}
\arguments{
\item{m}{One numeric value: a pre-estimated difference in means}

\item{se}{One numeric value: a pre-estimated standard error}

\item{df}{One numeric value: a pre-estimated degrees of freedom}
}
\value{
A list containing the following components:
\describe{
\item{p}{The p-value for the test.}
\item{std.err}{The estimated standard error of the difference in means.}
\item{m}{The observed difference in means between \code{x} and \code{y}.}
}
}
\description{
This function performs a composite t-test similar to \code{generalized.t.test}, but allows the user
to input precalculated values for the mean difference, standard error, and degrees of freedom.
}
