% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.tests.R
\name{generalized.ttest}
\alias{generalized.ttest}
\title{Generalized t-test for an Interval Null Hypothesis of Difference in Means}
\usage{
generalized.ttest(x, y, epsilon_l, epsilon_u, equal.var = F)
}
\arguments{
\item{x}{A non-empty numeric vector of data values}

\item{y}{A non-empty numeric vector of data values}

\item{epsilon_l}{A single numeric value representing the lower bound of the interval null hypothesis.}

\item{epsilon_u}{A single numeric value representing the lower bound of the interval null hypothesis.
This must be larger than or equal to \code{epsilon_l}.}

\item{equal.var}{A logical value indicating whether to assume equal variances between the two populations.
If \code{TRUE}, the function uses the pooled variance; if \code{FALSE} it uses the Welch approximation.
The default is \code{FALSE}}
}
\value{
A list containing the following components:
\describe{
\item{p}{The p-value for the test.}
\item{std.err}{The estimated standard error of the difference in means.}
\item{m}{The observed difference in means between \code{x} and \code{y}.}
}
}
\description{
Performs a generalized t-test to determine if the difference in means between
two numeric vectors \code{x} and \code{y} lies within the specified interval
[\code{epsilon_l}, \code{epsilon_u}]. This test is an extension of the standard t-test,
incorporating an interval null hypothesis
}
