% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.functions.R
\name{gm.mean}
\alias{gm.mean}
\title{Compute the Geometric Mean of a Numeric Vector}
\usage{
gm.mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector of positive values.}

\item{na.rm}{A logical indicating whether to remove all NA from \code{x}.}
}
\value{
The geometric mean of the values in \code{x}.
}
\description{
Compute the Geometric Mean of a Numeric Vector
}
