% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexa.test.R
\name{indexa.test}
\alias{indexa.test}
\title{Calculate Log-Fold Changes and Perform Interval Null Hypothesis Testing}
\usage{
indexa.test(indexa.obj, conds, epsilon.interval, equal.var = T)
}
\arguments{
\item{indexa.obj}{An object of class \code{indexa.mc}, as returned by the \code{indexa.mc} function.}

\item{conds}{A numeric or logical vector indicating the conditions. For a numeric vector,
values should be binary (1 for treatment or after condition, 0 for control or
before condition). For a logical vector, \code{TRUE} is equivalent to 1, and \code{FALSE}
is equivalent to 0.}

\item{epsilon.interval}{A numeric vector of length 2 specifying the interval bounds of the
interval assumption (see examples below). The first element is the
lower bound, and the second element is the upper bound.}

\item{equal.var}{Logical; if \code{TRUE}, the t-tests assume equal variances between the two
conditions. If \code{FALSE}, the Welch approximation is used.}
}
\value{
A \code{data.frame} containing the following columns:
\describe{
\item{\code{lfc.lower.bound}}{Lower bound of the median log fold change
across monte carlo samples.}
\item{\code{lfc.lower.bound}}{Upper bound of the median log fold change
across monte carlo samples.}
\item{\code{ctt.pval}}{Expected p-value for the composite t-test across all Monte
Carlo samples for each gene/microbe.}
\item{\code{cwt.pval}}{Expected p-value for the composite Wilcoxon rank test across
all Monte Carlo samples for each gene/microbe.}
\item{\code{gtt.pval}}{Expected p-value for the generalized t-test across all Monte
Carlo samples for each gene/microbe.}
\item{\code{ctt.pval.BH.adj}}{Expected Benjamini-Hochberg adjusted p-value for the
composite t-test across all Monte Carlo samples for
each gene/microbe.}
\item{\code{cwt.pval.BH.adj}}{Expected Benjamini-Hochberg adjusted p-value for the
composite Wilcoxon rank test across all Monte Carlo
samples for each gene/microbe.}
\item{\code{gtt.pval.BH.adj}}{Expected Benjamini-Hochberg adjusted p-value for the
generalized t-test across all Monte Carlo samples for
each gene/microbe.}
}
}
\description{
This function applies the CTT, GTT, and CWT tests in the simple case of
a comparison of two conditions.
}
\details{
#' This function computes the expected values (means) of p-values for various
statistical tests, including the composite t-test, generalized t-test, and
composite Wilcoxon rank test, across all Monte Carlo samples in the provided
\code{indexa.mc} object. It also computes the lower and upper bounds for the median
Log Fold Change (LFC) across all Monte Carlo Samples based on the interval
assumption given.
}
\note{
The median LFC is computed rather than the mean to accouny for potential
outliers in the distribution of LFCs (i.e., effect sizes) calculated across samples.
}
\examples{
# Example: Suppose a study compares the absolute abundances of microbes in the colons
# of individuals in a treatment group (indicated by `1`) vs. a control group (indicated
# by `0`). Suppose we expect total colonic microbial load (i.e., the scale) to be up to
# 3 fold higher in the treatment group. We can represent this as the interval assumption
# [0, log2(2)] in our analysis. As long as this interval assumption is correct, our
# absolute differential abundance analysis will control the false positive rate.

# Made-up dataset for colon study
Y <- rbind(c(4,0,1,9), c(5,1,9,0))
colnames(Y) <- c("control_1", "control_2", "treatment_1", "treatment_2")
rownames(Y) <- c("microbe_1", "microbe_2")

## Conditions: 0=control, 1=treatment
conds <- c(0, 0, 1, 1)

## Interval assumption (from example above)
epsilon.interval <- c(0, log2(3))

# Get the indexa object 
indexa.obj <- indexa.mc(Y, mc.samples=1000)

# Perform INDExA analysis
indexa.res <- indexa.test(indexa.obj, conds, epsilon.interval)

}
\author{
Kyle McGovern
}
