% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.functions.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Generate Random Samples from a Dirichlet Distribution}
\usage{
rdirichlet(n, alpha)
}
\arguments{
\item{n}{An integer specifying the number of samples to generate.}

\item{alpha}{A positive numeric vector representing the parameters of the Dirichlet distribution.}
}
\value{
A matrix with \code{n} rows and length of \code{alpha} columns, where each column
is a random sample drawn from the Dirichlet distribution.
}
\description{
Generate Random Samples from a Dirichlet Distribution
}
