% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.tests.R
\name{ttest.interval.fast}
\alias{ttest.interval.fast}
\title{Composite t-test for an Interval Null Hypothesis Difference in Means}
\usage{
ttest.interval.fast(x, y, epsilon_l, epsilon_u, equal.var = F)
}
\arguments{
\item{x}{A non-empty numeric vector of data values}

\item{y}{A non-empty numeric vector of data values}

\item{epsilon_l}{A single numeric value representing the lower bound of the interval null hypothesis.}

\item{epsilon_u}{A single numeric value representing the lower bound of the interval null hypothesis.
This must be larger than or equal to \code{epsilon_l}.}

\item{equal.var}{A logical value indicating whether to assume equal variances between the two populations.
If \code{TRUE}, the function uses the pooled variance; if \code{FALSE} it uses the Welch approximation.
The default is \code{FALSE}}
}
\value{
A list containing the following components:
\describe{
\item{p}{The p-value for the test.}
\item{std.err}{The estimated standard error of the difference in means.}
\item{m}{The observed difference in means between \code{x} and \code{y}.}
}
}
\description{
Performs a composite t-test to determine if the difference in means between
two numeric vectors \code{x} and \code{y} is greater than \code{epsilon_l} and less than \code{epsilon_u}.
The test involves performing two one-sided t-tests: one testing whether the difference
in means is less than \code{epsilon_l} and another testing if it is greater than \code{epsilon_u}.
If the p-values for each of these one-sided tests are \code{p_l} and \code{p_u}, respetively, then
the p-value for the composite t-test is calculated as \code{min(2*min(p_l,p_u), 1)}.
}
