% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.tests.R
\name{wilcox.interval.test}
\alias{wilcox.interval.test}
\title{Composite Wilcoxon Rank Test for Interval Null Hypothesis}
\usage{
wilcox.interval.test(x, y, epsilon_l, epsilon_u)
}
\arguments{
\item{x}{A non-empty numeric vector of data values}

\item{y}{A non-empty numeric vector of data values}

\item{epsilon_l}{A single numeric value representing the lower bound of the interval null hypothesis.}

\item{epsilon_u}{A single numeric value representing the lower bound of the interval null hypothesis.
This must be larger than or equal to \code{epsilon_l}.}

\item{equal.var}{A logical value indicating whether to assume equal variances between the two populations.
If \code{TRUE}, the function uses the pooled variance; if \code{FALSE} it uses the Welch approximation.
The default is \code{FALSE}}
}
\value{
A list containing the following components:
\describe{
\item{p}{The p-value for the test.}
\item{std.err}{The estimated standard error of the difference in means.}
\item{m}{The observed difference in means between \code{x} and \code{y}.}
}
}
\description{
Performs a composite Wilcoxon rank test to assess whether the difference between two
numeric vectors, \code{x} and \code{y}, lies within a specified interval. Specifically, the test
determines if the difference is greater than \code{epsilon_l} and less than \code{epsilon_u}.
}
\details{
This is done by conducting two one-sided Wilcoxon rank-sum tests: one to test if the
difference is less than \code{epsilon_l} and another to test if it is greater than \code{epsilon_u}.
The p-values from these tests are denoted as \code{p_l} and \code{p_u}, respectively. The p-value
for the composite test is then computed as \code{min(2 * min(p_l, p_u), 1)}.
}
